fix: PostgreSQL ROUND(double precision) compatibility
All ROUND() calls now use explicit ::numeric cast to avoid:
  function round(double precision, integer) does not exist

Affected files:
- run_analysis.py: all SQL queries (baseline, features, filters, repeat)
- 02_baseline_metrics.sql: PERCENTILE_CONT, AVG, SUM/ divisions
- 03_r_distribution.sql: COUNT/SUM division
- 04_winner_loser_features.sql: AVG over feature columns
- 05_repeat_signal_analysis.sql: SUM/COUNT division
- 06_candidate_filters.sql: header comment

Added regression test:
- tests/test_analysis_pg_compat.py: 19 tests scanning all SQL for unsafe ROUND()

Production code: unchanged
Scanner/Paper/Live: unaffected

+++ scripts/analysis/me_reverse_long_v1_v2/03_r_distribution.sql
@@ -1,5 +1,7 @@
 -- 03_r_distribution.sql
 -- R-distribution buckets for V1 and V2
+-- NOTE: All ROUND() calls use ::numeric cast to avoid
+--       PostgreSQL "function round(double precision, integer) does not exist"
 
 -- V1 R distribution
 SELECT
@@ -12,7 +14,7 @@ SELECT
         WHEN pnl_r >= 1.0 THEN '>= +1R'
     END AS r_bucket,
     COUNT(*) AS count,
-    ROUND(COUNT(*)::numeric / (SELECT COUNT(*) FROM dds.paper_trade WHERE scanner_name = 'MOMENTUM_EXHAUSTION_REVERSE_LONG_V1' AND status = 'CLOSED') * 100, 1) AS pct
+    ROUND((COUNT(*)::numeric / (SELECT COUNT(*) FROM dds.paper_trade WHERE scanner_name = 'MOMENTUM_EXHAUSTION_REVERSE_LONG_V1' AND status = 'CLOSED') * 100)::numeric, 1) AS pct
 FROM dds.paper_trade pt
 WHERE pt.scanner_name = 'MOMENTUM_EXHAUSTION_REVERSE_LONG_V1'
   AND pt.direction = 'LONG'
@@ -33,7 +35,7 @@ SELECT
         WHEN pnl_r >= 1.0 THEN '>= +1R'
     END AS r_bucket,
     COUNT(*) AS count,
-    ROUND(COUNT(*)::numeric / (SELECT COUNT(*) FROM dds.paper_trade WHERE scanner_name = 'MOMENTUM_EXHAUSTION_REVERSE_LONG_V2' AND status = 'CLOSED') * 100, 1) AS pct
+    ROUND((COUNT(*)::numeric / (SELECT COUNT(*) FROM dds.paper_trade WHERE scanner_name = 'MOMENTUM_EXHAUSTION_REVERSE_LONG_V2' AND status = 'CLOSED') * 100)::numeric, 1) AS pct
 FROM dds.paper_trade pt
 WHERE pt.scanner_name = 'MOMENTUM_EXHAUSTION_REVERSE_LONG_V2'
   AND pt.direction = 'LONG'
